+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 
 *2017-01-18*
 
-Initial commit.
+Initial commit
 
 
 1.0.1
@@ -13,3 +13,19 @@ Initial commit.
 
 - Fixed module exposure (using `ngc`)
 - Removed unnecessary `ElementRef` inclusion
+
+
+1.0.2
+-----
+
+*2017-01-19*
+
+- Compile with `ngc` on NPM `prepublish` hook
+
+
+1.1.0
+-----
+
+*2017-01-31*
+
+- **BREAKING CHANGE:** renamed module to `ChecklistModule`